requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
paho-mqtt>=2.1.0
python-dotenv>=1.0.1
httpx>=0.27.0
pydantic>=2.7.0
